backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Deprecated==1.2.12
Django==3.0.8
django-cors-headers==3.7.0
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pytz==2021.1
requests==2.25.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
